# docs/reviews/project-category-only-evidence/current-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/project-categories": {
      "get": {
        "tags": [
          "Project Categories"
        ],
        "summary": "List Project Categories",
        "description": "Full collection, createdAt ASC/id ASC. No query parameters, cursor, search or manual ordering. Creation cap 100 never truncates existing data.",
        "operationId": "list_2",
        "responses": {
          "200": {
            "description": "Full owned collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED or CSRF_INVALID for authenticated mutations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Project Categories"
        ],
        "summary": "Create a Project Category",
        "description": "Required key/body replays original 201 snapshot for 24 hours, even after rename/deletion. Raw name and field values define intent; normalization-equivalent raw names differ. No automatic defaults.",
        "operationId": "create_2",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1,
              "pattern": "[!-~]+"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created or original snapshot replayed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED or CSRF_INVALID for authenticated mutations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "CATEGORY_NAME_CONFLICT, IDEMPOTENCY_KEY_REUSED or REVISION_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "QUOTA_EXCEEDED: maximum 100 persisted Categories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v1/project-categories/{id}": {
      "get": {
        "tags": [
          "Project Categories"
        ],
        "summary": "Read a Project Category",
        "operationId": "get_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current owned Category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED or CSRF_INVALID for authenticated mutations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Project Categories"
        ],
        "summary": "Delete a Project Category",
        "description": "Any active or archived Project reference rejects deletion with CATEGORY_IN_USE. No cascading, clearing or reassignment. Repeated delete returns 404. No request body.",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9007199254740991,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED or CSRF_INVALID for authenticated mutations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "CATEGORY_IN_USE or REVISION_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Project Categories"
        ],
        "summary": "Rename a Project Category",
        "description": "Requires revision and name; same-value rename advances Category revision once, never Project revisions.",
        "operationId": "rename",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Renamed Category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "AUTH_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ACCOUNT_DISABLED or CSRF_INVALID for authenticated mutations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RESOURCE_NOT_FOUND: missing or inaccessible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "CATEGORY_NAME_CONFLICT or REVISION_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "fieldErrors",
          "message",
          "requestId"
        ]
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "Name: Java trim then NFC; case-sensitive workspace uniqueness; at most 100 normalized UTF-16 units. Unknown/duplicate/null fields rejected.",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "CategoryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1,
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "createdAt",
          "id",
          "name",
          "revision",
          "updatedAt"
        ]
      },
...
      "RenameCategoryRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "Both fields required. Same normalized name advances revision; revision-only PATCH is invalid. Name uses trim/NFC and case-sensitive workspace uniqueness.",
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64",
            "maximum": 9007199254740991,
            "minimum": 1
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          }
        },
        "required": [
          "name",
          "revision"
        ]
      },
...
      "CategoryListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Full collection size; creation admission limit 100 never truncates existing data",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
...
      "DeleteCategoryResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "deletedId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "deletedId"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}